Downgrade nginx dependency and refactor cookbook
The 2.5.0 of the nginx cookbook forces a depenndency on
version 2.0 of the build-essential and a version of the
yum and yum-epel cookbooks which conflict with the
postgresql and other cookbooks. The result is the Chef
server is unable to find a dependency resolution.

In Chef 11.4.x this seems to cause the chef-client to hang
indefinately while the chef-server is freaking out over
the impossible dependency resolution/conflicts.

Coderanger on #chef was kind enough to provide us a
means of identifying the issue:

https://gist.github.com/coderanger/9198932